# contracts/api/openapi.json
{
  "components": {
    "schemas": {
      "AiProviderProfileResponse": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "credential_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Reference"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "executable": {
            "title": "Executable",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "status_note": {
            "title": "Status Note",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider_id",
          "display_name",
          "base_url",
          "credential_reference",
          "enabled",
          "executable",
          "status_note",
          "created_at",
          "updated_at"
        ],
        "title": "AiProviderProfileResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local API for the AI skill learning platform.",
    "title": "云奕学 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/ai/provider-profiles": {
      "get": {
        "operationId": "list_ai_provider_profiles_ai_provider_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AiProviderProfileResponse"
                  },
                  "title": "Response List Ai Provider Profiles Ai Provider Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Ai Provider Profiles",
        "tags": [
          "ai-configuration"
        ]
      },
      "post": {
        "operationId": "create_ai_provider_profile_ai_provider_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAiProviderProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ai Provider Profile",
        "tags": [
          "ai-configuration"
        ]
      }
    },
...
    "/diagnostic-sessions": {
      "post": {
        "operationId": "create_diagnostic_session_diagnostic_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiagnosticSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/diagnostic-sessions/active": {
      "get": {
        "operationId": "get_active_diagnostic_session_diagnostic_sessions_active_get",
        "parameters": [
          {
            "in": "query",
            "name": "skill_id",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "skill_version",
            "required": true,
            "schema": {
              "maxLength": 50,
              "minLength": 1,
              "title": "Skill Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Active Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/diagnostic-sessions/latest": {
      "get": {
        "operationId": "get_latest_diagnostic_session_diagnostic_sessions_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "skill_id",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "skill_version",
            "required": true,
            "schema": {
              "maxLength": 50,
              "minLength": 1,
              "title": "Skill Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/diagnostic-sessions/{session_id}": {
      "get": {
        "operationId": "get_diagnostic_session_diagnostic_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/diagnostic-sessions/{session_id}/answers": {
      "post": {
        "operationId": "submit_diagnostic_answer_diagnostic_sessions__session_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Diagnostic Answer",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/diagnostic-sessions/{session_id}/answers/{question_id}/corrections": {
      "post": {
        "operationId": "correct_diagnostic_answer_diagnostic_sessions__session_id__answers__question_id__corrections_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Diagnostic Answer",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/diagnostic-sessions/{session_id}/end": {
      "post": {
        "operationId": "end_diagnostic_session_diagnostic_sessions__session_id__end_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
...
  }
}